feat: add vuetify to webview - refactor: isFuzzerStarted check
