+++ .vscode/launch.json
@@ -42,7 +42,7 @@
 			"outFiles": [
 				"${workspaceFolder}/src/clients/vscode/dist/**/*.js"
 			],
-			"preLaunchTask": "vscode ext - npm run watch",
+			"preLaunchTask": "npm run build Vuejs",
 			"sourceMaps": true,
 			//important, if not breakpoint remains unbound
 			"sourceMapPathOverrides": {
