+++ src/clients/vscode/package.json
@@ -52,8 +52,6 @@
   "dependencies": {
     "@types/axios": "^0.14.0",
     "@types/ws": "^8.5.3",
-    "axios": "^1.1.3",
-    "reconnecting-websocket": "^4.4.0",
     "ws": "^8.10.0"
   }
 }
