+++ src/clients/vscode/src/webview/package.json
@@ -12,6 +12,7 @@
     "bootstrap-vue": "^2.23.1",
     "bootstrap-vue-3": "^0.3.12",
     "core-js": "^3.8.3",
+    "styled-components": "^5.3.6",
     "vue": "^3.2.13",
     "vue-class-component": "^8.0.0-0"
   },
